Split commons into core and util
