+++ src/main/java/seedu/address/MainApp.java
@@ -4,7 +4,7 @@ import com.google.common.eventbus.Subscribe;
 import javafx.application.Application;
 import javafx.application.Platform;
 import javafx.stage.Stage;
-import seedu.address.commons.Version;
+import seedu.address.commons.core.Version;
 import seedu.address.model.AddressBook;
 import seedu.address.ui.Ui;
 import seedu.address.events.EventManager;
@@ -12,8 +12,8 @@ import seedu.address.events.controller.ExitAppRequestEvent;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.storage.StorageManager;
-import seedu.address.commons.Config;
-import seedu.address.commons.LoggerManager;
+import seedu.address.commons.core.Config;
+import seedu.address.commons.core.LoggerManager;
 
 import java.util.Map;
 import java.util.logging.Logger;
