+++ src/main/java/seedu/address/commons/core/LoggerManager.java
@@ -1,4 +1,4 @@
-package seedu.address.commons;
+package seedu.address.commons.core;
 
 import java.io.IOException;
 import java.util.logging.*;
