+++ src/main/java/seedu/address/commons/util/CollectionUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.commons;
+package seedu.address.commons.util;
 
 import java.util.Collection;
 import java.util.HashSet;
