+++ src/main/java/seedu/address/commons/util/FileUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.commons;
+package seedu.address.commons.util;
 
 import java.io.File;
 import java.io.IOException;
