+++ src/main/java/seedu/address/commons/util/FxViewUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.commons;
+package seedu.address.commons.util;
 
 import javafx.scene.Node;
 import javafx.scene.layout.AnchorPane;
