+++ src/main/java/seedu/address/commons/util/StringUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.commons;
+package seedu.address.commons.util;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
