+++ src/main/java/seedu/address/logic/commands/ClearCommand.java
@@ -1,6 +1,6 @@
 package seedu.address.logic.commands;
 
-import seedu.address.commons.CollectionUtil;
+import seedu.address.commons.util.CollectionUtil;
 import seedu.address.model.AddressBook;
 
 /**
