+++ src/main/java/seedu/address/logic/parser/Parser.java
@@ -1,15 +1,15 @@
 package seedu.address.logic.parser;
 
 import seedu.address.logic.commands.*;
-import seedu.address.commons.StringUtil;
+import seedu.address.commons.util.StringUtil;
 import seedu.address.exceptions.IllegalValueException;
 
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import static seedu.address.commons.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.commons.Messages.MESSAGE_UNKNOWN_COMMAND;
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 
 /**
  * Parses user input.
