+++ src/main/java/seedu/address/model/ModelManager.java
@@ -1,8 +1,8 @@
 package seedu.address.model;
 
 import javafx.collections.transformation.FilteredList;
-import seedu.address.commons.LoggerManager;
-import seedu.address.commons.StringUtil;
+import seedu.address.commons.core.LoggerManager;
+import seedu.address.commons.util.StringUtil;
 import seedu.address.events.model.LocalModelChangedEvent;
 import seedu.address.main.ComponentManager;
 import seedu.address.model.person.Person;
