+++ src/main/java/seedu/address/model/person/Email.java
@@ -1,7 +1,7 @@
 package seedu.address.model.person;
 
 
-import seedu.address.commons.CollectionUtil;
+import seedu.address.commons.util.CollectionUtil;
 import seedu.address.exceptions.IllegalValueException;
 
 /**
