+++ src/main/java/seedu/address/model/person/Person.java
@@ -1,6 +1,6 @@
 package seedu.address.model.person;
 
-import seedu.address.commons.CollectionUtil;
+import seedu.address.commons.util.CollectionUtil;
 import seedu.address.model.tag.UniqueTagList;
 
 import java.util.Objects;
