+++ src/main/java/seedu/address/model/person/UniquePersonList.java
@@ -2,7 +2,7 @@ package seedu.address.model.person;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.commons.CollectionUtil;
+import seedu.address.commons.util.CollectionUtil;
 import seedu.address.exceptions.DuplicateDataException;
 
 import java.util.*;
