+++ src/main/java/seedu/address/model/tag/Tag.java
@@ -1,7 +1,7 @@
 package seedu.address.model.tag;
 
 
-import seedu.address.commons.CollectionUtil;
+import seedu.address.commons.util.CollectionUtil;
 import seedu.address.exceptions.IllegalValueException;
 
 /**
