+++ src/main/java/seedu/address/model/tag/UniqueTagList.java
@@ -2,7 +2,7 @@ package seedu.address.model.tag;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.commons.CollectionUtil;
+import seedu.address.commons.util.CollectionUtil;
 import seedu.address.exceptions.DuplicateDataException;
 
 import java.util.*;
