+++ src/main/java/seedu/address/storage/StorageManager.java
@@ -1,15 +1,15 @@
 package seedu.address.storage;
 
 import com.google.common.eventbus.Subscribe;
-import seedu.address.commons.FileUtil;
+import seedu.address.commons.util.FileUtil;
 import seedu.address.events.model.LocalModelChangedEvent;
 import seedu.address.events.storage.*;
 import seedu.address.exceptions.DataConversionException;
 import seedu.address.main.ComponentManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.commons.Config;
-import seedu.address.commons.LoggerManager;
+import seedu.address.commons.core.Config;
+import seedu.address.commons.core.LoggerManager;
 
 import java.io.File;
 import java.io.FileNotFoundException;
