+++ src/main/java/seedu/address/storage/XmlFileStorage.java
@@ -1,6 +1,6 @@
 package seedu.address.storage;
 
-import seedu.address.commons.XmlUtil;
+import seedu.address.commons.util.XmlUtil;
 import seedu.address.exceptions.DataConversionException;
 
 import javax.xml.bind.JAXBException;
