+++ src/main/java/seedu/address/ui/HelpWindow.java
@@ -5,8 +5,8 @@ import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.web.WebView;
 import javafx.stage.Stage;
-import seedu.address.commons.FxViewUtil;
-import seedu.address.commons.LoggerManager;
+import seedu.address.commons.util.FxViewUtil;
+import seedu.address.commons.core.LoggerManager;
 
 import java.util.logging.Logger;
