+++ src/main/java/seedu/address/ui/ResultDisplay.java
@@ -6,7 +6,7 @@ import javafx.scene.Node;
 import javafx.scene.control.TextArea;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
-import seedu.address.commons.FxViewUtil;
+import seedu.address.commons.util.FxViewUtil;
 
 /**
  * A controller for the status bar that is displayed at the header of the application.
