+++ src/main/java/seedu/address/ui/Ui.java
@@ -8,7 +8,7 @@ import javafx.scene.control.Alert.AlertType;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
 import seedu.address.MainApp;
-import seedu.address.commons.StringUtil;
+import seedu.address.commons.util.StringUtil;
 import seedu.address.events.EventManager;
 import seedu.address.events.controller.JumpToListRequestEvent;
 import seedu.address.events.controller.PersonPanelSelectionChangedEvent;
@@ -18,8 +18,8 @@ import seedu.address.events.controller.ShowHelpEvent;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.ReadOnlyPerson;
-import seedu.address.commons.Config;
-import seedu.address.commons.LoggerManager;
+import seedu.address.commons.core.Config;
+import seedu.address.commons.core.LoggerManager;
 
 import java.io.File;
 import java.util.logging.Logger;
