+++ src/test/java/guitests/AddCommandTest.java
@@ -3,7 +3,7 @@ package guitests;
 import guitests.guihandles.PersonCardHandle;
 import org.junit.Test;
 import seedu.address.logic.commands.AddCommand;
-import seedu.address.commons.Messages;
+import seedu.address.commons.core.Messages;
 import seedu.address.testutil.TestPerson;
 import seedu.address.testutil.TestUtil;
