+++ src/test/java/guitests/FindCommandTest.java
@@ -1,7 +1,7 @@
 package guitests;
 
 import org.junit.Test;
-import seedu.address.commons.Messages;
+import seedu.address.commons.core.Messages;
 import seedu.address.testutil.TestPerson;
 
 import static org.junit.Assert.assertTrue;
