+++ src/test/java/seedu/address/commons/util/JsonUtilTest.java
@@ -1,10 +1,8 @@
-package seedu.address.commons;
+package seedu.address.commons.util;
 
 import org.junit.Test;
 
 import java.io.IOException;
-
-import static org.junit.Assert.assertEquals;
 
 /**
  * Tests JSON Read and Write
