+++ src/test/java/seedu/address/commons/util/XmlUtilTest.java
@@ -1,4 +1,4 @@
-package seedu.address.commons;
+package seedu.address.commons.util;
 
 import org.junit.Rule;
 import org.junit.Test;
