+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -25,7 +25,7 @@ import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
-import static seedu.address.commons.Messages.*;
+import static seedu.address.commons.core.Messages.*;
 
 public class LogicManagerTest {
