+++ src/test/java/seedu/address/testutil/TestUtil.java
@@ -14,9 +14,9 @@ import junit.framework.AssertionFailedError;
 import org.loadui.testfx.GuiTest;
 import org.testfx.api.FxToolkit;
 import seedu.address.TestApp;
-import seedu.address.commons.FileUtil;
-import seedu.address.commons.OsDetector;
-import seedu.address.commons.XmlUtil;
+import seedu.address.commons.util.FileUtil;
+import seedu.address.commons.core.OsDetector;
+import seedu.address.commons.util.XmlUtil;
 import seedu.address.exceptions.IllegalValueException;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
